# use the full pacing rate, also when not in slow start