@ironsteel 2019-04-08
Cool :) I was also comparing this code with the code in chromium (for gQUIC) and didn't figure out why it was different. 
Does it make sense to cherry-pick this commit in the gQUIC branch?