@marten-seemann 2019-04-09
@ironsteel #1855 is probably the more important commit to cherry-pick.