@ironsteel 2019-04-09
@marten-seemann Oh I see, should I make a pull request to the gQUIC branch with these two commits?